+++ platform/src/sdk/models/shared/status.ts
@@ -39,6 +39,3 @@ export type Status = ClosedEnum<typeof Status>;
 /** @internal */
 export const Status$inboundSchema: z.ZodNativeEnum<typeof Status> = z
   .nativeEnum(Status);
-/** @internal */
-export const Status$outboundSchema: z.ZodNativeEnum<typeof Status> =
-  Status$inboundSchema;